requirements.txt
ib_async>=1.0.0
openai>=1.30.0
python-dotenv>=1.0.0
pandas>=2.2.0
pydantic>=2.7.0
pytest>=8.0.0
